backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
pandas>=2.2.2
faiss-cpu>=1.7.4
sentence-transformers>=3.0.1
Pillow>=10.3.0
numpy>=1.26.4
pytest>=8.2.0
